Consolidate database population functions
Updated manage.py to take the new file structure into account
